[Chore #160963333] implemented folder structure,status_code,validation and email
