+++ api/v2/models/__init__.py
@@ -1,0 +1,3 @@
+import sys
+sys.path.insert(0,r'C:\Users\Ronny\fast-food-fast')
+
